Move rename to the new test infra
