+++ src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Refactoring/RenameEndpointTest.cs
@@ -467,82 +467,6 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Refactoring.Test
             Assert.Null(result);
         }
 
-        [Fact]
-        public async Task Handle_Rename_SingleServer_CSharpEditsAreMapped()
-        {
-            var input = """
-                <div></div>
-
-                @{
-                    var $$myVariable = "Hello";
-
-                    var length = myVariable.Length;
-                }
-                """;
-
-            var newName = "newVar";
-
-            var expected = """
-                <div></div>
-
-                @{
-                    var newVar = "Hello";
-
-                    var length = newVar.Length;
-                }
-                """;
-
-            // Arrange
-            TestFileMarkupParser.GetPosition(input, out var output, out var cursorPosition);
-            var codeDocument = CreateCodeDocument(output);
-            var csharpSourceText = codeDocument.GetCSharpSourceText();
-            var csharpDocumentUri = new Uri("C:/path/to/file.razor__virtual.g.cs");
-            var serverCapabilities = new ServerCapabilities()
-            {
-                RenameProvider = true
-            };
-
-            // A null mapping service means we are exercising the workspace edit remapping code
-            await using var csharpServer = await CSharpTestLspServerHelpers.CreateCSharpLspServerAsync(csharpSourceText, csharpDocumentUri, serverCapabilities, razorSpanMappingService: null).ConfigureAwait(false);
-            await csharpServer.OpenDocumentAsync(csharpDocumentUri, csharpSourceText.ToString()).ConfigureAwait(false);
-
-            var razorFilePath = "C:/path/to/file.razor";
-            var documentContextFactory = new TestDocumentContextFactory(razorFilePath, codeDocument, version: 1337);
-            var languageServerFeatureOptions = Mock.Of<LanguageServerFeatureOptions>(options =>
-                options.SupportsFileManipulation == true &&
-                options.SingleServerSupport == true &&
-                options.CSharpVirtualDocumentSuffix == ".g.cs" &&
-                options.HtmlVirtualDocumentSuffix == ".g.html"
-                , MockBehavior.Strict);
-            var languageServer = new RenameLanguageServer(csharpServer, csharpDocumentUri);
-            var documentMappingService = new DefaultRazorDocumentMappingService(languageServerFeatureOptions, documentContextFactory, LoggerFactory);
-            var projectSnapshotManager = Mock.Of<ProjectSnapshotManagerBase>(p => p.Projects == new[] { Mock.Of<ProjectSnapshot>(MockBehavior.Strict) }, MockBehavior.Strict);
-            var projectSnapshotManagerAccessor = new TestProjectSnapshotManagerAccessor(projectSnapshotManager);
-            var projectSnapshotManagerDispatcher = new LSPProjectSnapshotManagerDispatcher(LoggerFactory);
-            var searchEngine = new DefaultRazorComponentSearchEngine(Dispatcher, projectSnapshotManagerAccessor, LoggerFactory);
-
-            var endpoint = new RenameEndpoint(projectSnapshotManagerDispatcher, documentContextFactory, searchEngine, projectSnapshotManagerAccessor, languageServerFeatureOptions, documentMappingService, languageServer, TestLoggerFactory.Instance);
-
-            codeDocument.GetSourceText().GetLineAndOffset(cursorPosition, out var line, out var offset);
-            var request = new RenameParamsBridge
-            {
-                TextDocument = new TextDocumentIdentifier
-                {
-                    Uri = new Uri(razorFilePath)
-                },
-                Position = new Position(line, offset),
-                NewName = newName
-            };
-
-            // Act
-            var result = await endpoint.Handle(request, CancellationToken.None);
-
-            // Assert
-            var edits = result.DocumentChanges.Value.First.FirstOrDefault().Edits.Select(e => e.AsTextChange(codeDocument.GetSourceText()));
-            var newText = codeDocument.GetSourceText().WithChanges(edits).ToString();
-            Assert.Equal(expected, newText);
-        }
-
         private static IEnumerable<TagHelperDescriptor> CreateRazorComponentTagHelperDescriptors(string assemblyName, string namespaceName, string tagName)
         {
             var fullyQualifiedName = $"{namespaceName}.{tagName}";
@@ -681,50 +605,6 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Refactoring.Test
 
             var endpoint = new RenameEndpoint(projectSnapshotManagerDispatcher, documentContextFactory, searchEngine, projectSnapshotManagerAccessor, languageServerFeatureOptions, documentMappingService, languageServer, TestLoggerFactory.Instance);
             return endpoint;
-        }
-
-        private class RenameLanguageServer : ClientNotifierServiceBase
-        {
-            private readonly CSharpTestLspServer _csharpServer;
-            private readonly Uri _csharpDocumentUri;
-
-            public RenameLanguageServer(CSharpTestLspServer csharpServer, Uri csharpDocumentUri)
-            {
-                _csharpServer = csharpServer;
-                _csharpDocumentUri = csharpDocumentUri;
-            }
-
-            public override OmniSharp.Extensions.LanguageServer.Protocol.Models.InitializeParams ClientSettings { get; }
-
-            public override Task OnStarted(ILanguageServer server, CancellationToken cancellationToken)
-            {
-                return Task.CompletedTask;
-            }
-
-            public override Task<IResponseRouterReturns> SendRequestAsync(string method)
-            {
-                throw new NotImplementedException();
-            }
-
-            public async override Task<IResponseRouterReturns> SendRequestAsync<T>(string method, T @params)
-            {
-                Assert.Equal(RazorLanguageServerCustomMessageTargets.RazorRenameEndpointName, method);
-                var renameParams = Assert.IsType<DelegatedRenameParams>(@params);
-
-                var renameRequest = new RenameParams()
-                {
-                    TextDocument = new TextDocumentIdentifier()
-                    {
-                        Uri = _csharpDocumentUri
-                    },
-                    Position = renameParams.ProjectedPosition,
-                    NewName = renameParams.NewName,
-                };
-
-                var result = await _csharpServer.ExecuteRequestAsync<RenameParams, WorkspaceEdit>(Methods.TextDocumentRenameName, renameRequest, CancellationToken.None);
-
-                return new TestResponseRouterReturn(result);
-            }
         }
     }
 }
